moved notis.jsx to components directory
